move DefaultTemplate() to the file with same name
